+++ backend/src/main/java/com/synergyhub/events/rbac/RolePermissionChangedEvent.java
@@ -9,12 +9,12 @@ import lombok.Getter;
 public class RolePermissionChangedEvent extends BaseEvent {
     private final Role role;
     private final String changeType; // "ASSIGNED" or "REVOKED"
-    private final java.util.Set<Integer> permissionIds;
+    private final java.util.Set<Long> permissionIds;
 
     public RolePermissionChangedEvent(
             Role role,
             String changeType,
-            java.util.Set<Integer> permissionIds,
+            java.util.Set<Long> permissionIds,
             User actor,
             String ipAddress) {
         super(actor, ipAddress);
