+++ backend/src/main/java/com/synergyhub/events/task/TaskAssignedEvent.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 
 @Getter
 public class TaskAssignedEvent extends TaskEvent {
-    private final Integer assigneeId;
+    private final Long assigneeId;
     private final String assigneeName;
 
-    public TaskAssignedEvent(User actor, String ipAddress, Integer projectId, String projectName, String taskTitle, Integer assigneeId, String assigneeName) {
+    public TaskAssignedEvent(User actor, String ipAddress, Long projectId, String projectName, String taskTitle, Long assigneeId, String assigneeName) {
         super(actor, ipAddress, projectId, projectName, taskTitle);
         this.assigneeId = assigneeId;
         this.assigneeName = assigneeName;
