+++ backend/src/main/java/com/synergyhub/events/task/TaskMovedEvent.java
@@ -9,7 +9,7 @@ public class TaskMovedEvent extends TaskEvent {
     private final String fromStatus;
     private final String toStatus;
 
-    public TaskMovedEvent(User actor, String ipAddress, Integer projectId, String projectName, String taskTitle, String fromStatus, String toStatus) {
+    public TaskMovedEvent(User actor, String ipAddress, Long projectId, String projectName, String taskTitle, String fromStatus, String toStatus) {
         super(actor, ipAddress, projectId, projectName, taskTitle);
         this.fromStatus = fromStatus;
         this.toStatus = toStatus;
