+++ backend/src/main/java/com/synergyhub/exception/RoleNotFoundException.java
@@ -2,7 +2,7 @@ package com.synergyhub.exception;
 
 public class RoleNotFoundException extends RuntimeException {
 
-    public RoleNotFoundException(Integer roleId) {
+    public RoleNotFoundException(Long roleId) {
         super("Role with ID " + roleId + " not found");
     }
