+++ backend/src/main/java/com/synergyhub/exception/UnauthorizedOrganizationAccessException.java
@@ -2,7 +2,7 @@ package com.synergyhub.exception;
 
 public class UnauthorizedOrganizationAccessException extends RuntimeException {
 
-    public UnauthorizedOrganizationAccessException(Integer organizationId, Integer userId) {
+    public UnauthorizedOrganizationAccessException(Long organizationId, Long userId) {
         super("User " + userId + " is not authorized to access organization " + organizationId);
     }
 
@@ -10,3 +10,4 @@ public class UnauthorizedOrganizationAccessException extends RuntimeException {
         super(message);
     }
 }
+
