+++ backend/src/main/java/com/synergyhub/exception/UnauthorizedProjectAccessException.java
@@ -2,7 +2,7 @@ package com.synergyhub.exception;
 
 public class UnauthorizedProjectAccessException extends RuntimeException {
 
-    public UnauthorizedProjectAccessException(Integer projectId, Integer userId) {
+    public UnauthorizedProjectAccessException(Long projectId, Long userId) {
         super("User " + userId + " is not authorized to access project " + projectId);
     }
