+++ backend/src/main/java/com/synergyhub/repository/SprintRepository.java
@@ -12,41 +12,41 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface SprintRepository extends JpaRepository<Sprint, Integer> {
+public interface SprintRepository extends JpaRepository<Sprint, Long> {
 
     // Find all sprints in a project
-    List<Sprint> findByProjectIdOrderByStartDateDesc(Integer projectId);
+    List<Sprint> findByProjectIdOrderByStartDateDesc(Long projectId);
 
     // Find active sprint in project
     @Query("SELECT s FROM Sprint s WHERE s.project.id = :projectId AND s.status = 'ACTIVE'")
-    Optional<Sprint> findActiveSprintByProjectId(@Param("projectId") Integer projectId);
+    Optional<Sprint> findActiveSprintByProjectId(@Param("projectId") Long projectId);
 
     // Find sprints by status
-    List<Sprint> findByProjectIdAndStatus(Integer projectId, SprintStatus status);
+    List<Sprint> findByProjectIdAndStatus(Long projectId, SprintStatus status);
 
     // Find upcoming sprints
     @Query("SELECT s FROM Sprint s WHERE s.project.id = :projectId AND s.startDate > :date AND s.status = 'PLANNING' ORDER BY s.startDate ASC")
-    List<Sprint> findUpcomingSprintsByProjectId(@Param("projectId") Integer projectId, @Param("date") LocalDate date);
+    List<Sprint> findUpcomingSprintsByProjectId(@Param("projectId") Long projectId, @Param("date") LocalDate date);
 
     // Find completed sprints
     @Query("SELECT s FROM Sprint s WHERE s.project.id = :projectId AND s.status = 'COMPLETED' ORDER BY s.endDate DESC")
-    List<Sprint> findCompletedSprintsByProjectId(@Param("projectId") Integer projectId);
+    List<Sprint> findCompletedSprintsByProjectId(@Param("projectId") Long projectId);
 
     // Check if sprint name exists in project
-    boolean existsByProjectIdAndName(Integer projectId, String name);
+    boolean existsByProjectIdAndName(Long projectId, String name);
 
     // Find sprint with tasks
-    @Query("SELECT s FROM Sprint s LEFT JOIN FETCH s.tasks WHERE s.id = :id")
-    Optional<Sprint> findByIdWithTasks(@Param("id") Integer id);
+        @Query("SELECT s FROM Sprint s LEFT JOIN FETCH s.tasks WHERE s.id = :id")
+        Optional<Sprint> findByIdWithTasks(@Param("id") Long id);
 
     // Count sprints by project
-    long countByProjectId(Integer projectId);
+    long countByProjectId(Long projectId);
 
     // Find overlapping sprints
     @Query("SELECT s FROM Sprint s WHERE s.project.id = :projectId " +
             "AND s.status IN ('ACTIVE', 'PLANNING') " +
             "AND ((s.startDate <= :endDate AND s.endDate >= :startDate))")
-    List<Sprint> findOverlappingSprints(@Param("projectId") Integer projectId,
+    List<Sprint> findOverlappingSprints(@Param("projectId") Long projectId,
                                         @Param("startDate") LocalDate startDate,
                                         @Param("endDate") LocalDate endDate);
 }
