+++ backend/src/main/java/com/synergyhub/repository/SsoUserMappingRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SsoUserMappingRepository extends JpaRepository<SsoUserMapping, Integer> {
+public interface SsoUserMappingRepository extends JpaRepository<SsoUserMapping, Long> {
     
     Optional<SsoUserMapping> findByProviderAndExternalId(SsoProvider provider, String externalId);
