+++ backend/src/main/java/com/synergyhub/security/UserContext.java
@@ -3,17 +3,18 @@ package com.synergyhub.security;
 import java.util.Set;
 
 public class UserContext {
-    private final Integer id;
+    private final Long id;
     private final String email;
     private final Set<String> roles;
 
-    public UserContext(Integer id, String email, Set<String> roles) {
+    public UserContext(Long id, String email, Set<String> roles) {
         this.id = id;
         this.email = email;
         this.roles = roles;
     }
 
-    public Integer getId() { return id; }
+    public Long getId() { return id; }
     public String getEmail() { return email; }
     public Set<String> getRoles() { return roles; }
 }
+
